django_spinwheel/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
distlib==0.3.8
Django==5.0.2
django-adminplus==0.6
django-filter==24.2
django-object-actions==4.2.0
djangorestframework==3.15.1
filelock==3.13.4
html5lib==1.1
idna==3.6
lxml==5.1.0
Markdown==3.6
oscrypto==1.3.0
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
StrEnum==0.4.15
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
virtualenv==20.25.1
webencodings==0.5.1
xhtml2pdf==0.2.15
